+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+ - `PartialEq` and `Eq` are now derived for `Codec`, `Avc1`, and `Mp4a`.
+
 ## 0.2.0 - 2024-03-17
 
 ### Changed